requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
discord==2.3.2
discord.py==2.5.2
dnspython==2.7.0
emoji==2.14.1
frozenlist==1.6.0
idna==3.10
multidict==6.4.4
pillow==11.2.1
pillow_heif==0.22.0
propcache==0.3.1
pymongo==4.13.0
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
yarl==1.20.0
